% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi612-package.R
\docType{package}
\name{gi612-package}
\alias{gi612}
\alias{gi612-package}
\title{gi612: Data Package For GI 612 Course}
\description{
A package contain dataset required for GI 612--Introduction to GIS for coastal and Marine Resource Management offered at the Insitute of Marine Sciences of the University of Dar es Salaam.
}
\author{
\strong{Maintainer}: Masumbuko Semba \email{lugosemba@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
