% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi612-package.R
\docType{data}
\name{spatial}
\alias{spatial}
\alias{africa}
\alias{africa_catch}
\alias{land}
\alias{lakes}
\alias{rivers}
\alias{reef}
\alias{simple_feature}
\alias{world_countries}
\alias{tracks}
\title{spatial data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 59 rows and 8 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3315 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1354 rows and 38 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1454 rows and 35 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1043 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 11 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 177 rows and 11 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1467 rows and 25 columns.
}
\usage{
africa

africa_catch

land

land

lakes

rivers

reef

simple_feature

world_countries

tracks
}
\description{
These datasets contain spatial data required to accomplish GI 612
}
\keyword{datasets}
