% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi612-package.R
\docType{data}
\name{tabular}
\alias{tabular}
\alias{algoa_ctd}
\alias{cast_locations}
\alias{drifter_confluence}
\alias{fao_catch_tz}
\alias{fao_catch_wio}
\alias{fao_catch_world}
\alias{octopus_data}
\alias{kimbiji_kizimkazi}
\alias{primary_productivity}
\alias{points}
\alias{sampling_points}
\alias{sst_anomaly}
\alias{tafiri_chl}
\alias{tafiri_pp}
\alias{tafiri_sst}
\title{Vertical Profile of CTD}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2796 rows and 12 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2377 rows and 14 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 69 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 690 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16353 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1079 rows and 10 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11 rows and 6 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8360 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 3 columns.
}
\usage{
algoa_ctd

cast_locations

drifter_confluence

fao_catch_tz

fao_catch_wio

fao_catch_world

octopus_data

kimbiji_kizimkazi

primary_productivity

points

sampling_points

sst_anomaly

tafiri_chl

tafiri_pp

tafiri_sst
}
\description{
A data frame containing vertical profile of CTD
}
\details{
Tabular data

These datasets contain tabular data required to accomplish GI 612
}
\author{
Masumbuko Semba
}
\keyword{datasets}
